requirements.txt
flask
flask-cors
pillow
numpy
gunicorn

# Tells pip to look for CPU-only versions of torch
--index-url https://download.pytorch.org/whl/cpu
torch
torchvision